Add padding to ClusterManager
When using GoogleMap with padding, the getVisibleRegion() returns a smaller region, than what can be seen through the padding. If a marker is in the padding area, it is no longer rendered, although it should still be visible. So this adds the ability to give the padding also to the Clustermanager to render all markers that should be visible.